requirements.txt
SQLAlchemy~=2.0.39
dotenv~=0.9.9
python-dotenv
psycopg2-binary
fastapi~=0.115.12
pydantic~=2.10.6
tronpy~=0.5.0
uvicorn


